Add casts to clean up warnings like	off_t passed as  unsigned int

+++ hfaxd/Status.c++
@@ -167,7 +167,7 @@ HylaFAXServer::getServerStatus(const char* fileName, fxStr& status)
         (void) Sys::fstat(fd, sb);
         status.resize((u_int) sb.st_size);
         char* buff = new char[sb.st_size];
-        int n = Sys::read(fd, buff, sb.st_size);
+        int n = Sys::read(fd, buff, (size_t) sb.st_size);
         status = buff;
         Sys::close(fd);
         if (n > 0 && status[n-1] == '\n') n--;
